Adjust Mini Cart
For your consideration...  I put the General,basket resource call inside the quantity check to avoid the duplicate "basket" effect.  I did drop the applied total because my sense is that the mini cart will be in a lot of headers.  The totals were taking up too much room for my liking but this  was a personal decision.   I think this is a more sensible approach to working with the 2 key value pairs.